refactor: move MetricCollector
