+++ src/utils/datadog.ts
@@ -1,5 +1,5 @@
 import agent from 'elastic-apm-node';
 
-import MetricCollector from '../MetricCollector';
+import MetricCollector from './MetricCollector';
 
 export const datadog = new MetricCollector(agent);
